Fix folder structure Fix imports Remove Cart unused component
