+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
-import logo from "logo.svg";
-import './App.css';
-import Products from "products.jsx";
+import logo from "./logo.svg";
+import "./App.css";
+import Products from "./products.jsx";
 
 function App() {
   return <Products></Products>;
